Use the pathlib2 when it's available
Now that pathlib is receiving improvements and changes in general, the
backward compatibility layer called `pathlib' was deprecated in favor
of a module with different name that can be installed in
parallel (pathlib2). This change favors importing pathlib2 since
that's most likely only installed in systems that actually require
that to function.